pyproject.toml
[project]
name = "primogem-auth"
version = "0.1.0"
description = "JWT + Ed25519 authentication library"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "el-li07", email = "radkevich@sfedu.ru" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
    "pydantic-settings>=2.5.0",
    "bcrypt>=4.1.0",
    "sqlalchemy>=2.0.0",
    "slowapi>=0.1.9",
    "requests>=2.31.0",
]

[project.scripts]
primogem-cli = "primogem.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["primogem*"]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"